Add BloomFilter64 to support large Bloom filters
Add BloomFilter64 that is like BloomFilter, but uses ulong to
represent m and count, and uses Buckets64 and Hash128.

Add Buckets64 that is like Buckets, but uses ulong to represent count
and uses multiple arrays to overcome the C# array size limit of
approximately 2^31 elements and the default .NET limit of 2^31 bytes
per object.

Add Hash128 that is like Hash, but provides 128 bits of hash material.
Also, add Hash128.ComputeHashAndSum() to improve performance by
avoiding the conversion of bytes to string, then string to bytes.

Add Utils.OptimalM64() that takes a ulong n and calculates a ulong m.

Add Utils.HashKernel128() and Utils.HashKernel128ReturnValue for
working with 128 bits of hash material.  In Utils.HashKernel128(),
avoid the use of Skip() and Take() to improve performance.

Add TestBloomFilter64.cs, TestBuckets64.cs, TestHash128.cs to test
the new classes.